Created two phrase class methods `checkLetter()`: and `showMatchedLetter()`:
 `checkLetter()`: Checks to see if the letter selected by the player matches a letter in the phrase.
showMatchedLetter()`: Reveals the letter(s) on the board that matches the
player's selection. To reveal the matching letter(s), select all of the letter DOM
elements that have a CSS class name that matches the selected letter and
replace each selected element's `hide` CSS class with the `show` CSS class.