Life lost and game over logic
`removeLife()`: This method removes a life from the scoreboard, by replacing
one of the `liveHeart.png` images with a `lostHeart.png` image (found in the
`images` folder) and increments the `missed` property. If the player has five
missed guesses (i.e they're out of lives), then end the game by calling the
`gameOver()` method.
`gameOver()`: This method displays the original start screen overlay, and
depending on the outcome of the game, updates the overlay `h1` element with
a friendly win or loss message, and replaces the overlay’s `start` CSS class with
either the `win` or `lose` CSS class.